Renamed latex_numeric_typo, removed Typo.js dependency
